requirements.txt
pandas
numpy
scikit-learn
psycopg2-binary
faker
fastapi
uvicorn
mlflow
evidently
